requirements.txt
Django==5.1.4
django-restframework==0.0.1
djangorestframework==3.15.2
httpx==0.28.1
asyncio==3.4.3
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1